move Notifications components out of Geral subfolder
