+++ website/src/Components/Admin/ConfigManager.js
@@ -27,8 +27,8 @@ import {
 } from "defaultOcrConfigs";
 
 import ReturnButton from 'Components/FileSystem/ReturnButton';
-import ConfirmLeave from 'Components/Notifications/Geral/ConfirmLeave';
-import Notification from 'Components/Notifications/Geral/Notification';
+import ConfirmLeave from 'Components/Notifications/ConfirmLeave';
+import Notification from 'Components/Notifications/Notification';
 import ConfirmActionPopup from 'Components/Form/ConfirmActionPopup';
 import CheckboxList from 'Components/Form/CheckboxList';
 import Footer from 'Components/Footer/Footer';
